requirements.txt
boto3==1.37.11
fastapi==0.115.11
mysql-connector-python==9.2.0
python-dotenv==1.0.1
reportlab==4.3.1
requests==2.32.3
uvicorn==0.34.0
